+++ src/util/pointer_expr.h
@@ -16,6 +16,131 @@ Author: Daniel Kroening, kroening@kroening.com
 
 class namespacet;
 
+/// The pointer type
+/// These are both 'bitvector_typet' (they have a width)
+/// and 'type_with_subtypet' (they have a subtype)
+class pointer_typet : public bitvector_typet
+{
+public:
+  pointer_typet(typet _subtype, std::size_t width)
+    : bitvector_typet(ID_pointer, width)
+  {
+    subtype().swap(_subtype);
+  }
+
+  signedbv_typet difference_type() const
+  {
+    return signedbv_typet(get_width());
+  }
+
+  static void check(
+    const typet &type,
+    const validation_modet vm = validation_modet::INVARIANT)
+  {
+    DATA_CHECK(vm, !type.get(ID_width).empty(), "pointer must have width");
+  }
+};
+
+/// Check whether a reference to a typet is a \ref pointer_typet.
+/// \param type: Source type.
+/// \return True if \p type is a \ref pointer_typet.
+template <>
+inline bool can_cast_type<pointer_typet>(const typet &type)
+{
+  return type.id() == ID_pointer;
+}
+
+/// \brief Cast a typet to a \ref pointer_typet
+///
+/// This is an unchecked conversion. \a type must be known to be \ref
+/// pointer_typet. Will fail with a precondition violation if type
+/// doesn't match.
+///
+/// \param type: Source type.
+/// \return Object of type \ref pointer_typet.
+inline const pointer_typet &to_pointer_type(const typet &type)
+{
+  PRECONDITION(can_cast_type<pointer_typet>(type));
+  pointer_typet::check(type);
+  return static_cast<const pointer_typet &>(type);
+}
+
+/// \copydoc to_pointer_type(const typet &)
+inline pointer_typet &to_pointer_type(typet &type)
+{
+  PRECONDITION(can_cast_type<pointer_typet>(type));
+  pointer_typet::check(type);
+  return static_cast<pointer_typet &>(type);
+}
+
+/// This method tests,
+/// if the given typet is a pointer of type void.
+inline bool is_void_pointer(const typet &type)
+{
+  return type.id() == ID_pointer && type.subtype().id() == ID_empty;
+}
+
+/// The reference type
+///
+/// Intends to model C++ reference. Comparing to \ref pointer_typet should
+/// never be null.
+class reference_typet : public pointer_typet
+{
+public:
+  reference_typet(typet _subtype, std::size_t _width)
+    : pointer_typet(std::move(_subtype), _width)
+  {
+    set(ID_C_reference, true);
+  }
+
+  static void check(
+    const typet &type,
+    const validation_modet vm = validation_modet::INVARIANT)
+  {
+    PRECONDITION(type.id() == ID_pointer);
+    const reference_typet &reference_type =
+      static_cast<const reference_typet &>(type);
+    DATA_CHECK(
+      vm, !reference_type.get(ID_width).empty(), "reference must have width");
+    DATA_CHECK(
+      vm, reference_type.get_width() > 0, "reference must have non-zero width");
+  }
+};
+
+/// Check whether a reference to a typet is a \ref reference_typet.
+/// \param type: Source type.
+/// \return True if \p type is a \ref reference_typet.
+template <>
+inline bool can_cast_type<reference_typet>(const typet &type)
+{
+  return can_cast_type<pointer_typet>(type) && type.get_bool(ID_C_reference);
+}
+
+/// \brief Cast a typet to a \ref reference_typet
+///
+/// This is an unchecked conversion. \a type must be known to be \ref
+/// reference_typet. Will fail with a precondition violation if type
+/// doesn't match.
+///
+/// \param type: Source type.
+/// \return Object of type \ref reference_typet.
+inline const reference_typet &to_reference_type(const typet &type)
+{
+  PRECONDITION(can_cast_type<reference_typet>(type));
+  return static_cast<const reference_typet &>(type);
+}
+
+/// \copydoc to_reference_type(const typet &)
+inline reference_typet &to_reference_type(typet &type)
+{
+  PRECONDITION(can_cast_type<reference_typet>(type));
+  return static_cast<reference_typet &>(type);
+}
+
+bool is_reference(const typet &type);
+
+bool is_rvalue_reference(const typet &type);
+
 /// \brief Split an expression into a base object and a (byte) offset
 class object_descriptor_exprt : public binary_exprt
 {
